Split initial conditions functionality out
